refactor: start to split user input into a new module
