+++ Makefile
@@ -1,6 +1,6 @@
 CC=gcc
 CFLAGS=-Wall --std=c99 
-OBJECTS=./build/vector.o ./build/my_string.o ./build/node.o ./build/word_linked_list.o ./build/game_print.o
+OBJECTS=./build/vector.o ./build/my_string.o ./build/node.o ./build/word_linked_list.o ./build/game_print.o ./build/user_input.o
 
 main: ./src/main.c $(OBJECTS)
 	$(CC) $(CFLAGS) ./src/main.c $(OBJECTS) -o play_evil_hangman
@@ -18,6 +18,9 @@ main: ./src/main.c $(OBJECTS)
 	$(CC) $(CFLAGS) -c -o $@ $<
 
 ./build/game_print.o: ./src/ui/game_print.c
+	$(CC) $(CFLAGS) -c -o $@ $<
+
+./build/user_input.o: ./src/user_input/user_input.c
 	$(CC) $(CFLAGS) -c -o $@ $<
 
 clean:
