+++ src/user_input/user_input.h
@@ -1,0 +1,7 @@
+#include <stdio.h>
+
+int get_num_guesses_input(void);
+
+int get_word_length_input(void);
+
+int get_char_guess_input(void);
